// apps/mobile/lib/gen/model/evolutionStage.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Core API
 * Core API
 * OpenAPI spec version: 0.0.1
 */

export type EvolutionStage = typeof EvolutionStage[keyof typeof EvolutionStage];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EvolutionStage = {
  Baby: 'Baby',
  Young: 'Young',
  Adult: 'Adult',
  Prime: 'Prime',
  Legendary: 'Legendary',
} as const;